% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewiring.R
\name{addRandomEdges}
\alias{addRandomEdges}
\title{Add random edges to a supply network}
\usage{
addRandomEdges(
  g,
  n = 1,
  from = NULL,
  to = NULL,
  v2connect = NULL,
  multi.edges = FALSE,
  iie.ceil = 4
)
}
\arguments{
\item{g}{An igraph object.}

\item{n}{numeric; the number of edges to add.}

\item{from}{igraph.vs or character names; vertices sequence of possible edges
origins (default: NULL).}

\item{to}{igraph.vs or character names; vertices sequence of possible edges
origins (default: NULL).}

\item{v2connect}{igraph.vs or character names; vertices sequence edges origin
or destinations must be attached to (default: NULL).}

\item{multi.edges}{boolean; wheter or not multiple edges can be created
(default: FALSE).}

\item{iie.ceil}{numeric; maximal number of edges between intermediaries
in the graph (default: 4). TODO Must be changed to Inf}
}
\value{
An igraph object with \code{n} more edges.
}
\description{
Add random edges to a supply network
}
\examples{
## Generate a theoretical supply network, define producers (P),
## intermediaries (I) and distributors (D) vertices.
library(igraph)
g = make_tree(10)
V(g)[!degree(g, mode="in")]$type = "P"
V(g)[degree(g, mode="in") & degree(g, mode="in")]$type = "I"
V(g)[!degree(g, mode="out")]$type = "D"
V(g)$color = c("red","green","blue")[factor(V(g)$type)]
plot(g)
## Add 5 edges going from P to D (short supply chains)
g2 = addRandomEdges(g, n=5, from=V(g)[type=="P"], to=V(g)[type=="D"])
plot(g2)
## Add 3 edges connected to I (belonging long supply chains)
g2 = addRandomEdges(g, n=3, v2connect=V(g)[type=="I"])

}
