% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewiring_tools.R
\name{adjustAvgDegree}
\alias{adjustAvgDegree}
\title{Adjust the average degree of a graph}
\usage{
adjustAvgDegree(g, avg.degree = 1.470968, ...)
}
\arguments{
\item{g}{an igraph object.}

\item{avg.degree}{numeric; the expected average degree.}

\item{...}{other arguments passed to \code{\link{addRandomEdges}} to
constrain the add of edges. For example, if we want the edges to be added
to long supply chains, we pass \code{v2connect=V(g)[type=="I"]}}
}
\value{
an igraph object.
}
\description{
During a rewiring process, several operations such as nodes deletion
can make change the average degree of a graph. This function aims to
re-adjust the number of edges to a given level by adding or deleting edges
randomly.
}
\examples{
library(igraph)
g = generateSupplyNet()
plot(g)
ecount(g) / vcount(g)
g = adjustAvgDegree(g, avg.degree=3)
ecount(g) / vcount(g)
plot(g)

}
