% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateVertices.R
\name{aggregateVertices}
\alias{aggregateVertices}
\title{Aggregate vertices of a graph}
\usage{
aggregateVertices(g, by, v.agg = V(g))
}
\arguments{
\item{g}{An igraph object}

\item{by}{Vector of the same length than vertices to aggregate them.}

\item{v.agg}{igraph.vs or character; the vertices to aggregate. Default is
all graph vertices.}
}
\value{
An igraph object.
}
\description{
Aggregate vertices of an igraph object. For vertices numeric columns, mean of
groups is computed. For character class columns, uniques values are pasted.
}
\examples{
library(igraph)
g = erdos.renyi.game(100, 120, 'gnm')
plot(g)
by = sample(letters[1:10], length(V(g)), replace=TRUE)
g2 = aggregateVertices(g, by)
plot(g2)

}
