% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRaster.R
\name{createRaster}
\alias{createRaster}
\title{Create fitted rasters}
\usage{
createRaster(extent, xref = NULL, yref = NULL, scale = 1)
}
\arguments{
\item{extent}{sf or sp; spatial object defining the extent covered by raster.
Can be points, lines or polygons.}

\item{xref}{numeric; x coordinate of reference pixel bottomleft corner.
If NULL (default), the center of the extent is taken.}

\item{yref}{numeric; y coordinate of reference pixel bottomleft corner.
If NULL (default), the center of the extent is taken.}

\item{scale}{numeric; resolution of the raster cells, expressed in units of
\code{extent} CRS. If CRS is EPSG:2154, scale is in meters. If EPSG:4326,
it is in decimal degrees.}
}
\value{
a RasterLayer object with same CRS than \code{extent}.
}
\description{
Creates a raster covering a spatial object \code{extent}.
Resolution \code{scale} is passed in the units of \code{extent} object.
Raster pixels can be fitted on a reference point passed with
\code{xref} and \code{yref}.
}
\examples{
\donttest{
# Create rasters covering French Seine river basin, fitted on "Point Zero"
# of french roads in Paris city center.
library(sf)
library(spData)
library(raster)
data(seine, package="spData")
pt_zero = st_sf(st_sfc(st_point(c(652215.52, 6861681.77)), crs=2154))
coords = st_coordinates(pt_zero)
rast = createRaster(extent=seine, xref=coords[1], yref=coords[2], scale=100)
values(rast) = rnorm(ncell(rast))
plot(rast)
plot(st_geometry(seine), col="blue", add=TRUE)
plot(st_geometry(pt_zero), col="red", add=TRUE)
# Test with sp class in WGS84 CRS, without specifying xref and yref
library(sp)
data(SplashDams, package="spData")
proj4string(SplashDams) = CRS("+proj=longlat +datum=WGS84")
rast = createRaster(SplashDams, scale=0.0001)
values(rast) = rnorm(ncell(rast))
plot(rast)
plot(SplashDams, col="blue", add=TRUE)
}

}
