% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagDensity.R
\name{dagDensity}
\alias{dagDensity}
\title{Density of a Directed Acyclic Graph (DAG)}
\usage{
dagDensity(g)
}
\arguments{
\item{g}{an igraph object; a directed acycic graph, with is_dag=TRUE.}
}
\value{
a numeric between 0 and 1.
}
\description{
A Directed Acyclic Graph (DAG) is a hierarchical structure of 3 levels of
vertices: the source set, the intermediaries set and the targets set. Given
that edges outgoing from targets and incoming to sources cannot exist, this
metric compute a relevant graph density for DAGs?
}
\examples{
library(igraph)
g = make_tree(10)
edge_density(g)
edge_density(as.undirected(g))
is_dag(g)
dagDensity(g)

}
