% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewiring.R
\name{deleteMultipleEdges}
\alias{deleteMultipleEdges}
\title{Delete multiple edges of a graph}
\usage{
deleteMultipleEdges(g, verbose = FALSE)
}
\arguments{
\item{g}{An igraph object}

\item{verbose}{Wether or not to print a message if multiple edges are deleted.
(default: FALSE).}
}
\value{
An igraph object without any multiple edges.
}
\description{
Delete multiple edges in the graph if any, and print message if wanted.
Messages can be useful to monitor an iterative rewiring process for example.
}
\examples{
library(igraph)
g = generateSupplyNet()
vij = sample(V(g), 2)
g = add_edges(g, c(vij, vij))
g = deleteMultipleEdges(g, verbose=TRUE)

}
