% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency.R
\name{efficiencyRe}
\alias{efficiencyRe}
\title{Supply network reachability-efficiency}
\usage{
efficiencyRe(g)
}
\arguments{
\item{g}{igraph object.}
}
\value{
A numeric between 0 and 1.
}
\description{
Divide the reachability of the network by the effort spent by producers and
distributors, i.e. their average degree.
}
\examples{
library(igraph)
g = make_tree(10)
efficiencyRe(g)

}
