% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSupplyNet.R
\name{generateSupplyNet}
\alias{generateSupplyNet}
\title{Generate theoretical supply network}
\usage{
generateSupplyNet(P = 67, I = 18, D = 70, E = 228)
}
\arguments{
\item{P}{numeric; number of producers (default: 67).}

\item{I}{numeric; number of intermediaries (default: 18).}

\item{D}{numeric; number of distributors (default: 70).}

\item{E}{numeric; number of edges (default: 228).}
}
\value{
igraph object; a supply network.
}
\description{
Generate a theorectical supply network composed of three types of vertices:
producers, intermediaries, and distributors. Output graph is a Directed
Acyclic Graph (DAG) in which every producers and intermediaries has >= 1
customer, and every intermediaries and distributors have a supplier.
}
\examples{
net = generateSupplyNet()
plot(net)

}
