% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBridges.R
\name{getBridges}
\alias{getBridges}
\title{Get bridges edges of a graph}
\usage{
getBridges(g)
}
\arguments{
\item{g}{An igraph object}
}
\value{
an edge sequence of the graph defined as bridges.
}
\description{
An edge in an undirected connected graph is a bridge iff removing it disconnects
the graph. For a disconnected undirected graph, definition is similar, a bridge
is an edge removing which increases number of disconnected components.
Function is available in igraph C version, but not in R.
\href{https://igraph.discourse.group/t/function-to-find-edges-which-are-bridges-in-r-igraph/154}{Solution}
proposed by Szabolcs Horvát.
}
\examples{
library(igraph)
set.seed(523)
g = erdos.renyi.game(100, 120, 'gnm')
bridges = getBridges(g)
E(g)$color = "gray"
E(g)[bridges]$color = "red"
plot(g, vertex.size=6, vertex.label=NA)

}
