% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph2sf.R
\name{igraph2sf}
\alias{igraph2sf}
\title{Creating spatial vertices or edges from igraph object}
\usage{
igraph2sf(g, what = c("vertices", "edges"), coords = c("x", "y"), crs = 2154)
}
\arguments{
\item{g}{An igraph object}

\item{what}{Character wether to return spatial "vertices" or spatial "edges. Default is "vertices"}

\item{coords}{names of the numeric columns holding coordinates in vertices dataframe}

\item{crs}{Coordinate reference system: integer with the EPSG code, or character with proj4string}
}
\value{
a sf class object: point if type = "vertices", linestring if "edges".
}
\description{
Extends \code{\link[igraph]{get.data.frame}} by adding spatial dimension with \code{\link[sf]{st_as_sf}}
}
\examples{
\donttest{
library(igraph)
library(sf)
data(USairports, package="igraphdata")
g = USairports
url = 'https://raw.githubusercontent.com/jpatokal/openflights/master/data/airports.dat'
apts = read.csv(url, header=FALSE)
V(g)$lng = apts[match(V(g)$name, apts[,5]), 8]
V(g)$lat = apts[match(V(g)$name, apts[,5]), 7]
g = delete.vertices(g, which(is.na(V(g)$lat)))
sf_e = igraph2sf(g, what="e", coords=c("lng", "lat"), crs=4326)
sf_v = igraph2sf(g, what="v", coords=c("lng", "lat"), crs=4326)
plot(st_geometry(sf_e[sample(seq(nrow(sf_e), 50)), ]), col="blue")
plot(st_geometry(sf_v), col="red", add=TRUE)
}

}
