% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewiring_tools.R
\name{lengthenShortEdge}
\alias{lengthenShortEdge}
\title{Turn a short supply chain into a long supply chain}
\usage{
lengthenShortEdge(g)
}
\arguments{
\item{g}{an igraph object.}
}
\value{
an igraph object.
}
\description{
Choose a random short supply-chain edge going directly from producers to
distributors, delete it, and replace it by a path of two edges length passing
through an intermediary chosen randomly.
}
\examples{
## Turn all short SC into long SC
library(igraph)
g = generateSupplyNet(E=200)
while (length(E(g)[E(g)$type=="PD"]))
  g = lengthenShortEdge(g)
ecount(g)

}
