% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiSapply.R
\name{multiSapply}
\alias{multiSapply}
\title{multiSapply function takes a vector as first argument and next one can specify
multiple functions that are to be applied to this vector.
\href{Solution}{https://www.r-bloggers.com/applying-multiple-functions-to-data-frame/}
proposed by Bogumił Kamiński}
\usage{
multiSapply(...)
}
\arguments{
\item{...}{first element is the vector or list to apply the function on.
Folliwing arguments are the functions.}
}
\value{
a data.frame with as many columns as functions passed.
}
\description{
Apply Multiple function to an object
}
\examples{
data(cars)
multiSapply(cars, min, mean, max)

}
