% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{nbSupplyChains}
\alias{nbSupplyChains}
\title{Overall number of supply chains in a supply network}
\usage{
nbSupplyChains(g)
}
\arguments{
\item{g}{igraph object; the supply network.}
}
\value{
a positive integer.
}
\description{
A supply chain is a path starting from a producer and ending to a distributor.
A supply network counts as many supply chains as connected producer-distributors
pairs.
}
\examples{
library(igraph)
g = generateSupplyNet()
nbSupplyChains(g)

}
