% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertices_local_indices.R
\name{neighborsDegree}
\alias{neighborsDegree}
\title{Summarize degree sequence of a vertex neighbors}
\usage{
neighborsDegree(
  g,
  vs = V(g),
  ngh.mode = c("all", "in", "out"),
  deg.mode = c("all", "in", "out"),
  FUN = min
)
}
\arguments{
\item{g}{igraph object.}

\item{vs}{igraph.vs; vertices sequence of which the local index will be
calculated (default: all vertices).}

\item{ngh.mode}{character; the neighbors mode, either 'all', 'in', 'out'
(default: 'all').}

\item{deg.mode}{character; the degree mode, either 'all', 'in', 'out'
(default: 'all').}

\item{FUN}{function used to summarize the degree sequence (default: min).}
}
\value{
A numeric vector of the same length than vertices sequence vs.
}
\description{
For a given vertex, get its neighbors degree sequence, and summarize it with
a function. Can be used to identify a kind of articulations points: if the
minimal degree of a vertex neighbors is > 1, its deletion does not isolate
another one.
}
\examples{
library(igraph)
set.seed(123)
g = erdos.renyi.game(10, 20, 'gnm', directed=TRUE)
plot(g)
neighborsDegree(g, FUN=max)

}
