% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency.R
\name{reachability}
\alias{reachability}
\title{Supply network reachability}
\usage{
reachability(g)
}
\arguments{
\item{g}{igraph object.}
}
\value{
A numeric between 0 and 1.
}
\description{
Reachability of a supply network measures the probability that a source
node (a producer) is connected to a target node through a path
It is equal to the number of supply paths from the sources set S to sink
set T, divided by the overall number of possible paths (|S| * |T|).
}
\examples{
## In a tree, reachability is equal to 1
library(igraph)
g = make_tree(10)
reachability(g)
## Theoretical 1 hub-centralized supply network
#TODO

}
