% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{relativeDegree}
\alias{relativeDegree}
\title{Vertex relative degree}
\usage{
relativeDegree(g, vs = V(g), mode = c("in", "out"))
}
\arguments{
\item{g}{igraph object.}

\item{vs}{igraph.vs; vertices sequence of which the local index will be
calculated (default: all vertices).}

\item{mode}{either 'in' our 'out': the relative degree mode (default: 'in').}
}
\value{
A numeric vector of the same length than vertices sequence vs.
}
\description{
Relative in- or out-degree of a vertex measures the balance between its
number of incoming and outgoing edges.
}
\examples{
library(igraph)
set.seed(123)
g = erdos.renyi.game(10, 20, 'gnm', directed=TRUE)
relativeDegree(g)

}
