% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewireEdge.R
\name{rewireEdge}
\alias{rewireEdge}
\title{Rewire origin or destination of an edge}
\usage{
rewireEdge(
  g,
  edge = sample(E(g), 1),
  what = c("origin", "destination"),
  vs = NULL,
  multi.edges = FALSE,
  ...
)
}
\arguments{
\item{g}{An igraph object}

\item{edge}{An igraph.es edge to rewire. Default is 1 random edge.}

\item{what}{For directed graphs, what to rewire: "origin" or "destination" ?
Default is "origin".}

\item{vs}{igraph.vs vertices towards which edge origin or destination can be rewired.}

\item{multi.edges}{Wheter or not multi edges can be created during the process.
Default is FALSE.}

\item{...}{Other parameters passed to \code{\link{getVertices}} function.}
}
\value{
An igraph object with 1 edge rewired.
}
\description{
Preferential attachment can be included by passing some paramters to ...
}
\examples{
library(igraph)
g = erdos.renyi.game(100, 120, 'gnm')
for (i in 1:100)
  g = rewireEdge(g)

}
