% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{shortEdgesRate}
\alias{shortEdgesRate}
\title{Short supply chains edges rate}
\usage{
shortEdgesRate(g)
}
\arguments{
\item{g}{an igraph object; the supply network.}
}
\value{
a numeric between 0 and 1.
}
\description{
Short supply chains are edges linking directly producers to distributors.
This function compute the rate of short SCs edges in the overall edges number.
}
\examples{
library(igraph)
g = generateSupplyNet()
shortEdgesRate(g)

}
