% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{simulatedAnnealing}
\alias{simulatedAnnealing}
\title{Graph Rewiring Simulated Annealing Improvement Optimization Algorithm}
\usage{
simulatedAnnealing(
  g,
  rewiring.FUN = NULL,
  fun2min = NULL,
  niter = Inf,
  temp0 = 1,
  cooling.method = c("exp", "fast", "boltz"),
  cooling.factor = 0.99,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{g}{an igraph object.}

\item{rewiring.FUN}{function of rewiring to apply to the graph at each time step.}

\item{fun2min}{function, metric to minimize.}

\item{niter}{maximal number of iterations. Default is Inf.}

\item{temp0}{initial temperature. Default is 1.}

\item{cooling.method}{method used to progressively decrease temperature. Defaut is "exp".}

\item{cooling.factor}{decimal between 0 and 1, used if cooling.method="exp".
Lower cooling factor make temperature (and candidate acceptation probability)
decreasing faster. Default is 0.95.}

\item{verbose}{Wether or not displaying fun2min evolution. Default is FALSE.}

\item{...}{additionnal arguments passed to rewiring.FUN}
}
\value{
a list of 7 elements:
\describe{
  \item{g}{the optimized graph}
  \item{value}{the fun2min value obtained at the end of the process}
  \item{g_list}{list of igraphs object. History of successive improvements}
  \item{iter}{number of iterations elapsed}
  \item{acceptation_rate}{Acceptation rate during the process}
  \item{temp0}{initial temperature passed as argument}
  \item{cooling.method}{cooling method passed as argument}
}
}
\description{
Try to minimize a given graph metric (function) through  rewiring process.
Metaheuristic trying to find a global optimum rather than local optimum.
}
\examples{
library(igraph)
g = sample_pa(n=100, power=1, m=1,  directed=FALSE)
is_dag(g)
res = simulatedAnnealing(g, rewiring.FUN=rewire, with=each_edge(prob=1),
 fun2min=transitivity, niter=10)

}
