% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{supplyTransitivity}
\alias{supplyTransitivity}
\title{Vertex supply transitivity (for intermediaires)}
\usage{
supplyTransitivity(g, vs = V(g))
}
\arguments{
\item{g}{igraph object.}

\item{vs}{igraph.vs; vertices sequence of which the local index will be
calculated (default: all vertices).}
}
\value{
numeric vector of the same length than vertices sequence vs.
}
\description{
Customized version of local transitivity index, adapted to intermediaries of a
supply network. Original transitivity measures the probability that the
adjacent vertices of a vertex are connected. This is sometimes also called
the clustering coefficient. This function measures the probability that the
in-neighbors (suppliers) and the out-neighbors (customers) of a vertex (an
intermediary) are connected. It reveals how an intermediary can be
"by-passed", "short-circuited" by parallel links, and thus useless.
}
\examples{
##  Supply Transitivity of in termediary reaches 1 after a transitive closure
library(igraph)
set.seed(123)
g = make_tree(5)
lay = layout_nicely(g)
plot(g, layout=lay)
supplyTransitivity(g, V(g)[2])
# Make transitive closure
m = floydAlgo(g)
m[m>1] = 1 # To adjacency matrix
g = graph_from_adjacency_matrix(m)
plot(g, layout=lay)
supplyTransitivity(g, V(g)[2])

}
\seealso{
\code{\link[igraph]{transitivity}}
}
