% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swapEdges.R
\name{swapEdges}
\alias{swapEdges}
\title{Swap two edges of a igraph object.}
\usage{
swapEdges(
  g,
  es = E(g),
  e1 = NULL,
  edges.types = c("random", "same", "different"),
  multi.edges = FALSE
)
}
\arguments{
\item{g}{An igraph object}

\item{es}{igraph.es edge sequence among which to choose edges to swap.
Default is all the graph edges.}

\item{e1}{one imposed edge to swap with another. Default is NULL.}

\item{edges.types}{Of what type must be the two swapped edges ? Can be "random",
"same" or "different". Default is "random". Require an E(g)$type attribute.}

\item{multi.edges}{whether or no multiple edges creation is allowed during
the process. Default is FALSE.}
}
\value{
an igraph object with 2 swapped edges.
}
\description{
Extends \code{\link[igraph]{rewire}} used with \code{\link[igraph]{keeping_degseq}}
with tunable parameters. Possible to force the swap of two edges of different
types, or of the same type. Edges swapping can be used to realize a Degree
Preserving Randomization (DPR). One of the two edges can be imposed (e1).
}
\examples{
## Generate a theoretical supply network, define producers (P), intermediaries (I)
## and distributors (D) vertices.
library(igraph)
g = make_tree(10)
V(g)[!degree(g, mode="in")]$type = "P"
V(g)[degree(g, mode="in") & degree(g, mode="out")]$type = "I"
V(g)[!degree(g, mode="out")]$type = "D"
V(g)$color = c("green","red","yellow")[factor(V(g)$type, levels=c("P","I","D"))]
plot(g)
## Swap edges to conserve supply from P to D.
for (i in 1:100) {
  i = ends(g, E(g), names=FALSE)[,1]
  j = ends(g, E(g), names=FALSE)[,2]
  E(g)$type = paste0(V(g)[i]$type, V(g)[j]$type)
  g = swapEdges(g, edges.types="diff")
  if (length(E(g)[which_multiple(g, E(g))]))
     stop("Multiedges")
  if (length(E(g)[which_loop(g, E(g))]))
     stop("Loops")
  if (length(V(g)[degree(g)==0]))
     stop("Isolated vertex")
}
plot(g)

}
