% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewiring.R
\name{updateGraphAttributes}
\alias{updateGraphAttributes}
\title{Update vertices and attributes of a graph}
\usage{
updateGraphAttributes(g)
}
\arguments{
\item{g}{an igraph object; the supply network.}
}
\value{
an igraph object.
}
\description{
First, different degrees (deg, indeg, outdeg) are computed and attached as
vertices attributes. Then, given these degrees, vertice type (P, I, D) is
deducted. Edges type is equal to the concatenation of ends vertices types.
}
\examples{
library(igraph)
g = make_tree(10)
g = updateGraphAttributes(g)
V(g)$type
E(g)$type

}
